# openapi/atlas-api.yaml
openapi: 3.1.0
info:
  title: ATLAS API
  description: |
    API for ATLAS (Automated Tracking & Licensing Application System)
    Manages permit applications, documents, users, and audit logs.
  version: 1.0.0
  contact:
    name: ATLAS Development Team
    email: dev@atlas.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:5001
    description: Development server
  - url: https://api.atlas.example.com
    description: Production server
security:
  - EntraID: ['User.Read']
...
paths:
...
  /api/applications/{applicationId}/documents:
    post:
      tags: [Documents]
      summary: Upload document
      description: Upload a document for an application
      operationId: uploadDocument
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadDocumentRequest'
      responses:
        '200':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Citizen', 'Officer', 'Admin']
  /api/documents/{documentId}/download:
    get:
      tags: [Documents]
      summary: Download document
      description: Download a document from Azure Blob Storage
      operationId: downloadDocument
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Document ID
      responses:
        '200':
          description: Document file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Citizen', 'Officer', 'Admin']
  /api/permittypes:
    get:
      tags: [PermitTypes]
      summary: Get all permit types
      description: Retrieve list of permit types, optionally including inactive ones
      operationId: getPermitTypes
      parameters:
        - name: includeInactive
          in: query
          schema:
            type: boolean
            default: false
          description: Include inactive permit types in results
      responses:
        '200':
          description: List of permit types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermitTypeDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
    
    post:
      tags: [PermitTypes]
      summary: Create permit type
      description: Create a new permit type
      operationId: createPermitType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermitTypeRequest'
      responses:
        '201':
          description: Permit type created successfully
          content:
            application/json:
              schema:
                type: string
                format: uuid
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
  /api/permittypes/{id}:
    get:
      tags: [PermitTypes]
      summary: Get permit type by ID
      description: Retrieve details of a specific permit type
      operationId: getPermitTypeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Permit type ID
      responses:
        '200':
          description: Permit type details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermitTypeDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
    
    put:
      tags: [PermitTypes]
      summary: Update permit type
      description: Update an existing permit type
      operationId: updatePermitType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Permit type ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePermitTypeRequest'
      responses:
        '200':
          description: Permit type updated successfully
          content:
            application/json:
              schema:
                type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
    
    delete:
      tags: [PermitTypes]
      summary: Deactivate permit type
      description: Deactivate a permit type (soft delete)
      operationId: deactivatePermitType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Permit type ID
      responses:
        '200':
          description: Permit type deactivated successfully
          content:
            application/json:
              schema:
                type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
  /api/users:
    get:
      tags: [Users]
      summary: Get all users
      description: Retrieve list of users with optional role filtering
      operationId: getUsers
      parameters:
        - name: role
          in: query
          schema:
            type: string
          description: Filter by user role
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
    
    post:
      tags: [Users]
      summary: Create user account
      description: Register a new user account
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: string
                format: uuid
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security: []
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Retrieve details of a specific user
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
  /api/users/{id}/role:
    put:
      tags: [Users]
      summary: Update user role
      description: Update the role of an existing user
      operationId: updateUserRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '200':
          description: User role updated successfully
          content:
            application/json:
              schema:
                type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
...
components:
  schemas:
...
    PermitTypeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Business Permit'
        description:
          type: string
          example: 'Permit for business operations'
        fee:
          type: number
          format: decimal
          example: 150.00
        isActive:
          type: boolean
          default: true
      required: [id, name, description, fee, isActive]
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          example: 'user@example.com'
        firstName:
          type: string
          example: 'John'
        lastName:
          type: string
          example: 'Doe'
        role:
          type: string
          example: 'Citizen'
        department:
          type: string
          nullable: true
          example: 'Permits Office'
        isActive:
          type: boolean
          default: true
      required: [id, email, firstName, lastName, role, isActive]
...
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
      required: [title, status]
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
          required: [errors]
...
    UploadDocumentRequest:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        fileName:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
          format: int64
        blobUrl:
          type: string
          format: uri
        uploadedById:
          type: string        
          format: uuid
      required: [applicationId, fileName, contentType, fileSize]
    CreatePermitTypeRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        fee:
          type: number
          format: decimal        
      required: [name, description, fee]
    UpdatePermitTypeRequest:
      type: object
      properties:
        permitTypeId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        fee:
          type: number
          format: decimal
        isActive:
          type: boolean
        estimatedProcessingDays:
          type: integer
          format: int32
        deactivatedByAdminId:
          type: string
          format: uuid
          nullable: true
      required: [permitTypeId]
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        department:
          type: string
          nullable: true        
      required: [email, firstName, lastName, role]
    UpdateUserRoleRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string        
      required: [userId, role]
  responses:
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'
  securitySchemes:
    EntraID:
      type: oauth2
      description: Microsoft Entra ID (Azure AD) authentication
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/token
          scopes:
            User.Read: Read user profile
            Citizen: Citizen role access
            Officer: Officer role access
            Admin: Admin role access